Remove common file so we can just have unix and windows file. I do *not* want to try to decompose stuff in TdsParser to remove LocalDBAPI references on non-windows systems.
